Move GenericNode subclasses to generic.py, including Hsy
