+++ nodes/simple.py
@@ -181,159 +181,7 @@ class SimpleNode(Node):
         return df
 
 
-# class Oper
-
-class GenericNode(SimpleNode):
-    explanation = _(
-        """
-        GenericNode: A highly configurable node that processes inputs through a sequence of operations.
-
-        Operations are defined in the 'operations' parameter and executed in order.
-        Each operation works on its corresponding basket of nodes.
-        """
-    )
-    allowed_parameters = [
-        *SimpleNode.allowed_parameters,
-        StringParameter(local_id='operations', label='Comma-separated list of operations to execute in order')
-    ]
-    # Class-level default operations
-    DEFAULT_OPERATIONS = 'multiply,add,other,apply_multiplier'
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # Instance-level operations (can be overridden by subclasses)
-        self.default_operations = self.DEFAULT_OPERATIONS
-
-        # Operation registry
-        self.OPERATIONS: dict[str, Callable[..., tuple[Any, Any]]]  = {
-            'multiply': self._operation_multiply,
-            'add': self._operation_add,
-            'other': self._operation_other,
-            'apply_multiplier': self._operation_apply_multiplier,
-        }
-
-    def _get_input_baskets(self, nodes: list[Node]) -> dict[str, list[Node]]:
-        """Return a dictionary of node 'baskets' categorized by type."""
-        baskets: dict[str, list[Node]] = {
-            'additive': [],
-            'multiplicative': [],
-            'other': []
-        }
-
-        # Categorize nodes by tags first
-        for edge in self.edges:
-            if edge.output_node != self or edge.input_node not in nodes:
-                continue
-
-            node = edge.input_node
-            if 'additive' in edge.tags or 'additive' in node.tags:
-                baskets['additive'].append(node)
-            elif 'non_additive' in edge.tags or 'non_additive' in node.tags:
-                baskets['multiplicative'].append(node)
-            elif 'other_node' in edge.tags or 'other_node' in node.tags:
-                baskets['other'].append(node)
-            elif self.is_compatible_unit(self.unit, node.unit):
-                baskets['additive'].append(node)
-            else:
-                baskets['multiplicative'].append(node)
-
-        return baskets
-
-    # Operation wrapper functions
-    def _operation_multiply(self, df: ppl.PathsDataFrame | None, baskets: dict, **kwargs) -> tuple:
-        """Multiply all nodes in the multiplicative basket."""
-        nodes = baskets['multiplicative']
-        if nodes and len(nodes) > 0:
-            df = self.multiply_nodes_pl(
-                df=df,
-                nodes=nodes,
-                metric=kwargs.get('metric'),
-                unit=kwargs.get('unit'),
-                start_from_year=kwargs.get('start_from_year')
-            )
-            baskets['multiplicative'] = []
-        return df, baskets
-
-    def _operation_add(self, df: ppl.PathsDataFrame | None, baskets: dict, **kwargs) -> tuple:
-        """Add all nodes in the additive basket."""
-        nodes = baskets['additive']
-        if nodes and len(nodes) > 0:
-            df = self.add_nodes_pl(
-                df=df,
-                nodes=nodes,
-                metric=kwargs.get('metric'),
-                keep_nodes=kwargs.get('keep_nodes', False),
-                node_multipliers=kwargs.get('node_multipliers'),
-                unit=kwargs.get('unit'),
-                start_from_year=kwargs.get('start_from_year'),
-                ignore_unit=True
-            )
-            baskets['additive'] = []
-        return df, baskets
-
-    def _operation_other(self, df: ppl.PathsDataFrame | None, baskets: dict, **kwargs) -> tuple:
-        """Process other nodes - to be implemented by subclasses."""
-        if type(self) is GenericNode and len(baskets['other']) > 0:
-            raise NodeError(self, f"Generic node {self.id} cannot handle 'other' nodes.")
-        return df, baskets
-
-    def _operation_apply_multiplier(self, df: ppl.PathsDataFrame | None, baskets: dict, **kwargs) -> tuple:
-        """Apply the node's multiplier parameter to the dataframe."""
-        if df is None:
-            raise NodeError(self, "Cannot apply multiplier because no PathsDataFrame is available.")
-        mult = self.get_parameter_value('multiplier', required=False, units=True)
-        if mult:
-            df = df.multiply_quantity(VALUE_COLUMN, mult)
-        return df, baskets
-
-    def compute(self) -> ppl.PathsDataFrame:
-        """Process inputs according to the operations sequence."""
-        # Get operation sequence from parameter or class default
-        operations_str = self.get_parameter_value_str('operations', required=False) or self.default_operations
-        operations = [op.strip() for op in operations_str.split(',')]
-
-        # Get input dataset and categorize nodes
-        df = self.get_input_dataset_pl(tag='baseline', required=False)
-        baskets = self._get_input_baskets(self.input_nodes)
-
-        # Track original node counts for validation
-        original_node_counts = {basket: len(nodes) for basket, nodes in baskets.items()}
-
-        kwargs = {
-            'metric': None,
-            'unit': self.unit,
-            'start_from_year': None,
-            'keep_nodes': False,
-            'node_multipliers': None
-        }
-
-        # Apply operations in sequence
-        for op_name in operations:
-            if op_name not in self.OPERATIONS:
-                raise NodeError(self, f"Unknown operation: {op_name}")
-
-            operation_func = self.OPERATIONS[op_name]
-            df, baskets = operation_func(df, baskets, **kwargs)
-        if not isinstance(df, ppl.PathsDataFrame):
-            raise NodeError(self, "The output is not a PathsDataFrame.")
-
-        # Validate that all nodes were used
-        unused_nodes = [
-            (node.id, basket)
-            for basket, nodes in baskets.items()
-            if nodes and original_node_counts.get(basket, 0) > 0
-            for node in nodes]
-
-        if unused_nodes:
-            unused_str = ", ".join([f"{node_id} ({basket})" for node_id, basket in unused_nodes])
-            raise NodeError(self, f"Unused input nodes found after all operations: {unused_str}")
-
-        df = df.ensure_unit(VALUE_COLUMN, self.unit)
-
-        return df
-
-
-class AdditiveNode(GenericNode):
+class AdditiveNode(SimpleNode):
     explanation = _("""This is an Additive Node. It performs a simple addition of inputs.
 Missing values are assumed to be zero.""")
     allowed_parameters = [
@@ -1033,242 +881,6 @@ class FillNewCategoryNode2(AdditiveNode): # FIXME Merge into FillNewCategoryNode
                 df = df.filter(~pl.col(col).is_null())
             df = df.paths.to_narrow()
         return df
-
-
-class LeverNode(GenericNode):
-    explanation = _(
-        """LeverNode replaces the upstream computation completely, if the lever is enabled."""
-    )
-    allowed_parameters = [
-        *GenericNode.allowed_parameters,
-        StringParameter(local_id='new_category'),
-    ]
-
-    def _operation_override_with_lever(self, df: ppl.PathsDataFrame, baskets: dict, **kwargs) -> tuple:
-        """Override upstream computation with lever values if enabled."""
-        if df is None:
-            return None, baskets
-
-        # Get the lever node from the "other" basket
-        lever_nodes = baskets.get('other', [])
-        if not lever_nodes:
-            raise NodeError(self, "LeverNode requires an 'other_node' tagged input as lever")
-
-        lever = lever_nodes[0]  # Use the first lever node
-
-        if not isinstance(lever, ActionNode):
-            raise NodeError(self, f"Lever {lever} must be an action.")
-
-        # If lever is not enabled, return original dataframe
-        if not lever.is_enabled():
-            return df, baskets
-
-        # Get lever output and override values
-        dfl = lever.get_output_pl(target_node=self)
-        dfl = dfl.ensure_unit(VALUE_COLUMN, df.get_unit(VALUE_COLUMN))
-        out = df.paths.join_over_index(dfl, how='left', index_from='left')
-        out = out.with_columns(
-            (pl.when(pl.col(FORECAST_COLUMN))
-            .then(pl.col(VALUE_COLUMN + '_right'))
-            .otherwise(pl.col(VALUE_COLUMN))).alias(VALUE_COLUMN)
-        )
-        out = out.drop(VALUE_COLUMN + '_right')
-
-        if len(df) != len(out):
-            s = f"({len(out)} rows) as the affected node {self.id} ({len(df)} rows)."
-            raise NodeError(self, f"Lever {lever.id} must result in the same structure {s}")
-
-        return out, baskets
-
-    def _operation_fill_new_category(self, df: ppl.PathsDataFrame, baskets: dict, **kwargs) -> tuple:
-        """Fill in a new category with complement values."""
-        if df is None:
-            return None, baskets
-
-        category = self.get_parameter_value_str('new_category', required=True)
-        dim, cat = category.split(':')
-
-        df = df.ensure_unit(VALUE_COLUMN, 'dimensionless')
-
-        df2 = df.paths.sum_over_dims(dim)
-        df2 = df2.with_columns((pl.lit(1.0) - pl.col(VALUE_COLUMN)).alias(VALUE_COLUMN))
-        df2 = df2.with_columns(pl.lit(cat).cast(pl.Categorical).alias(dim))
-        df2 = df2.select(df.columns)
-
-        df = df.paths.concat_vertical(df2)
-        df = df.ensure_unit(VALUE_COLUMN, self.unit)
-
-        if self.get_parameter_value('drop_nans', required=False):  # FIXME Not consistent with the parameter name!
-            df = df.paths.to_wide()
-            for col in df.metric_cols:
-                df = df.filter(~pl.col(col).is_null())
-            df = df.paths.to_narrow()
-
-        return df, baskets
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # Register custom operations
-        self.OPERATIONS['override_with_lever'] = self._operation_override_with_lever
-        self.OPERATIONS['fill_new_category'] = self._operation_fill_new_category
-        # Set default operations sequence
-        self.default_operations = 'multiply,add,other,override_with_lever,fill_new_category,apply_multiplier'
-
-
-class WeightedSumNode(GenericNode):
-    explanation = _(
-        """
-        WeightedSumNode: Combines additive inputs using weights from a multidimensional weights DataFrame.
-        """
-    )
-    # Class-level default overrides parent
-    DEFAULT_OPERATIONS = 'multiply,add_with_weights,add,other,apply_multiplier'
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # Register the custom operation
-        self.OPERATIONS['add_with_weights'] = self._operation_add_with_weights
-
-    def _process_single_weighted_node(self, node: Node, node_weights: ppl.PathsDataFrame) -> ppl.PathsDataFrame:
-        """Process a single node with its weights and return the weighted output."""
-        # Get node output
-        node_output = node.get_output_pl(target_node=self)
-
-        # Find the metric column that has non-null values for this node
-        valid_metric = None
-        metrics = node_weights.metric_cols.copy()
-        for col in metrics:
-            if col in node_weights.columns and node_weights[col].null_count() < len(node_weights):
-                valid_metric = col
-                metrics.remove(col)
-                break
-
-        node_weights = node_weights.drop(metrics)
-        if not valid_metric:
-            raise NodeError(self, f"No valid metric column found for weight dataset for node {node.id}")
-
-        # Create a version with this metric renamed to VALUE_COLUMN
-        if valid_metric != VALUE_COLUMN:
-            node_weights = node_weights.rename({valid_metric: VALUE_COLUMN})
-
-        # Multiply node output with weights
-        return node_output.paths.multiply_with_dims(node_weights, how='inner')
-
-    def _combine_weighted_outputs(self, weights_df: ppl.PathsDataFrame, additive_nodes: list[Node]) -> ppl.PathsDataFrame | None:
-        """Process and combine all weighted node outputs."""
-        # Create a lookup map for additive nodes
-        node_map = {node.id: node for node in additive_nodes}
-        result = None
-
-        # Process each unique node in the weights dataframe
-        for node_id in weights_df['node'].unique():
-            if node_id not in node_map:
-                self.logger.warning(f"Node {node_id} not found in additive nodes")
-                continue
-
-            # Get node and its weights
-            node = node_map[node_id]
-            node_weights = weights_df.filter(pl.col('node') == node_id).drop('node')
-
-            # Process the node
-            weighted_output = self._process_single_weighted_node(node, node_weights)
-
-            # Add to result (first time initializes, subsequent adds)
-            if result is None:
-                result = weighted_output
-            else:
-                result = result.paths.add_with_dims(weighted_output, how='outer')
-
-        return result
-
-    def _operation_add_with_weights(self, df: ppl.PathsDataFrame, baskets: dict, **kwargs) -> tuple:
-        """Combine additive node outputs weighted by values in a multidimensional weights DataFrame."""
-        # Get weights dataframe from specifically tagged input
-        weights_df = self.get_input_dataset_pl(tag='input_node_weights', required=False)
-        if weights_df is None:
-            return df, baskets
-
-        # Focus specifically on nodes in the additive basket
-        additive_nodes = baskets['additive']
-        if not additive_nodes:
-            raise NodeError(
-                self,
-                "If node contains weights, it must contain additive input nodes (typically with tag 'additive')."
-            )
-
-        # Process all weighted nodes
-        result = self._combine_weighted_outputs(weights_df, additive_nodes)
-
-        if result is None:
-            self.logger.warning(f"No matching nodes found in weights DataFrame for {self.id}")
-            return df, baskets
-
-        # Remove processed nodes from additive basket (to prevent double-counting)
-        processed_nodes = [node for node in additive_nodes if node.id in weights_df['node'].unique()]
-        for node in processed_nodes:
-            baskets['additive'].remove(node)
-
-        return result, baskets
-
-
-class LogitNode(WeightedSumNode):
-    explanation = _(
-        """
-        LogitNode gives a probability of event given a baseline and several determinants.
-
-        The baseline is given as a dataset of observed values. The determinants are linearly
-        related to the logit of the probability:
-        ln(y / (1 - y)) = a + sum_i(b_i * X_i,)
-        where y is the probability, a is baseline, X_i determinants and b_i coefficients.
-        The node expects that a comes from dataset and sum_i(b_i * X_i,) is given by the input nodes
-        when operated with the GenericNode compute(). The probability is calculated as
-        ln(y / (1 - y)) = b <=> y = 1 / (1 + exp(-b)).
-        """
-    )
-
-    def _operation_logit_transform(self, df: ppl.PathsDataFrame, baskets: dict, **kwargs) -> tuple:
-        """Apply logit transform to combine observations with weighted sum."""
-        if df is None:
-            return None, baskets
-
-        # Ensure our weighted sum is in dimensionless units
-        df = df.ensure_unit(VALUE_COLUMN, 'dimensionless')
-
-        # Get observations dataset
-        df_obs = self.get_input_dataset_pl(tag='observations', required=False)
-        if df_obs is None:
-            raise NodeError(self, f"LogitNode {self.id} must have one dataset for baseline values.")
-
-        # Validate and transform observations to logit space
-        df_obs = df_obs.ensure_unit(VALUE_COLUMN, 'dimensionless')
-        test = df_obs.with_columns(
-            (pl.lit(0.0) < pl.col(VALUE_COLUMN)) & (pl.col(VALUE_COLUMN) < pl.lit(1.0))
-        )['literal'].all()
-
-        if not test:
-            raise NodeError(self, f"All values in {self.id} must be between 0 and 1, exclusive.")
-
-        df_obs = df_obs.with_columns(
-            (pl.col(VALUE_COLUMN) / (pl.lit(1.0) - pl.col(VALUE_COLUMN))).log().alias(VALUE_COLUMN)
-        )
-
-        # Join observations with weighted sum
-        df = df.paths.join_over_index(df_obs, how='outer', index_from='left')
-        df = df.sum_cols([VALUE_COLUMN, VALUE_COLUMN + '_right'], VALUE_COLUMN).drop(VALUE_COLUMN + '_right')
-
-        # Apply inverse logit function to get probabilities
-        expr = pl.lit(1.0) / (pl.lit(1.0) + (pl.lit(-1.0) * pl.col(VALUE_COLUMN)).exp())
-        df = df.with_columns(expr.alias(VALUE_COLUMN))
-        df = df.ensure_unit(VALUE_COLUMN, self.unit)
-
-        return df, baskets
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # Register the logit transform operation
-        self.OPERATIONS['logit_transform'] = self._operation_logit_transform
-        # Set default operations sequence
-        self.default_operations = 'multiply,add_with_weights,add,logit_transform,other,apply_multiplier'
 
 
 class ChooseInputNode(AdditiveNode):
